+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Occupancy changes can selectively override recurring calendar event based states, with configurable list of states, defaulting to `disarmed`,`armed_home` and `armed_night`. So if the calendar says `disarmed` but you've left the house it will override automatically to `armed_away`.
   - Effects occupancy driven changes, which continue even if a calendar event in progress, and start of calendar event, which won't
   override an occupancy state
+### Events
+- An `autoarming` event is now fired on the Home Assistant bus
 ### International
 - Strings now translated into Italian, French, German and Japanese
 ### Other